requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.0
jinja2==3.1.2
python-dateutil==2.8.2
markdown==3.5.1
schedule==1.2.0
flask==2.3.3
python-dotenv==1.0.0
